+++ build_magic/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.0
 
+* Added the `--target` and `-t` options to execute specific config file stages by name.
+* Added the ability to run specific config file stages by name passed as an argument if the config file has a default name of build-magic.yaml.
+* Added a process spinner to long-running stage start up commands.
 * Added the sequence number and total commands to the output for each command.
 * Removed the newline character at the end of the output displayed when in verbose mode.
 * Changed the stage end status from COMPLETE to DONE.